Move init_asl_t1_trans_wf from reg to res.
